Added type parameters to SummarySave so that I could pull the common implementation of `save` up, and pushed the various field-specific overrides down int to the controller classes so the three `GrantValueSave`, `DeadlineSave` and `TitleSave` traits are now no longer needed.
